Make the output is a 64bit width m_axi interface. The input is still a axis interface. Then the design is inserted back an axi iic and axi uart. Final bitstream is packed again to obtain BOOT.BIN. It can boot on the board successfully and the IP can run on the board.

+++ mz_libcaer_OF/src/ABMOF/abmof.cpp
@@ -25,7 +25,7 @@ static uint64_t imgNum = 0;
 static bool initSocketFlg = false;
 static uint16_t retSocket;
 
-static int32_t *eventSlice = (int32_t *)sds_alloc(DVS_HEIGHT * DVS_WIDTH);
+static uint64_t *eventSlice = (uint64_t *)sds_alloc(DVS_HEIGHT * DVS_WIDTH);
 
 // To trigger the tcp to send event slice
 static bool sendFlg = false;
@@ -926,7 +926,7 @@ int abmof(std::shared_ptr<const libcaer::events::PolarityEventPacket> polarityPk
     // Make suer sds_alloc allocate a right memory for eventSlice.
 	if(eventSlice == NULL)
 	{
-		eventSlice = (int32_t *)sds_alloc(DVS_HEIGHT * DVS_WIDTH);
+		eventSlice = (uint64_t *)sds_alloc(DVS_HEIGHT * DVS_WIDTH);
 		return retSocket;
 	}
 	if(eventsArraySize >= eventThreshold)
